refactor(xtask): move every test module into its own file
Eight modules kept their tests inline; three of them (ci/jobs, macos/bundle,
macos/bundle/identity) were already extracted by hand. This finishes the crate,
so `foo.rs` is what a module does and `foo/tests.rs` is what pins it —
the shape 26 files elsewhere in the workspace already use.

The `#[cfg(test)]` moves to the declaration, which is what carries
`clippy.toml`'s unwrap/expect exemption into the file: a fixture helper outside
any `#[test]` fn still gets it, and clippy `-D warnings` is green on the
result. Test count is unchanged at 44.

One trap the compiler caught and the README now states: `include_str!`
resolves relative to the file, so a path that reached the repo root from
`linux/package.rs` needs one more `..` from `linux/package/tests.rs`.
